requirements1.txt
websocket-client>=1.2.1
TimeScale
datetime
psycopg2>=2.9.1
